Update README to suggest the appropriate groups in Gemfile
This change makes it easier to understand which group factory_girl_rails belongs to in the Gemfile. By explicitly mentioning :development and :test we ensure that beginners won't add the gem into the global Gemfile namespace and therefore won't load this in production for no reason.

By the looks of it factory_girl_rails is in fact useful in :development as well to register itself for fixture replacements.